reorganize the chargepoint controller part of the repository
* remove the 'ev/' directory; we can add it back in later if it proves
  to be a useful grouping
* move the python code into a 'py/' subdirectory
* move the golang code into a more appropriately named 'api/'
  subdirectory
