+++ controller/chargepoint/api/schema/clear_shed_state.go
