+++ controller/chargepoint/api/schema/common.go
