+++ controller/chargepoint/api/schema/get_load.go
