+++ controller/chargepoint/api/schema/get_station_groups.go
