+++ controller/chargepoint/api/schema/get_stations.go
