+++ controller/chargepoint/api/schema/shed_load.go
