+++ controller/chargepoint/api/soap.go
@@ -1,4 +1,4 @@
-package chargepoint
+package api
 
 import (
 	"bytes"
