+++ controller/chargepoint/api/soap_test.go
@@ -1,4 +1,4 @@
-package chargepoint
+package api
 
 import (
 	"bytes"
