+++ setup.cfg
@@ -19,10 +19,10 @@ package_dir = kinney=
 packages =
     kinney
     kinney.controller
-    kinney.controller.ev
-    kinney.controller.ev.chargepoint
-    kinney.controller.ev.chargepoint.classes
-    kinney.controller.ev.chargepoint.test
+    kinney.controller.chargepoint
+    kinney.controller.chargepoint.py
+    kinney.controller.chargepoint.py.classes
+    kinney.controller.chargepoint.py.test
     kinney.orchestrator
 
 # YAPF configuration for Python formatting.
